Move models and routes back to root... Let's minimize the size of this PR.
