+++ app.js
@@ -5,13 +5,13 @@ var express = require('express')
 	, fs = require('fs')
     , path = require('path')
     , mongoose = require('mongoose')
-    , routes = require('./lib/routes')
-    , pkg = require('./lib/routes/package')
+    , routes = require('./routes')
+    , pkg = require('./routes/package')
     , passport = require('passport')
-    , user = require('./lib/routes/user')
+    , user = require('./routes/user')
     , users = require('./lib/users')
     , oxy_auth = require('./lib/oxygen_auth')
-    , stats = require('./lib/routes/stats')
+    , stats = require('./routes/stats')
     , basic_auth = require('./lib/basic_auth')
     , error = require('./lib/error')
 	, stats_update = require('./lib/stats_update');
