+++ lib/basic_auth.js
@@ -1,6 +1,6 @@
 var passport = require('passport')
   , BasicStrategy = require('passport-http').BasicStrategy
-  , UserModel = require('../lib/models/user').UserModel;
+  , UserModel = require('../models/user').UserModel;
 
 ////////////////////////
 // Passport
