+++ lib/packages.js
@@ -1,10 +1,10 @@
-var PackageModel = require('../lib/models/package').PackageModel
+var PackageModel = require('../models/package').PackageModel
 	, error = require('./error')
 	, users = require('./users')
 	, async = require('async')
 	, _ = require('underscore')
 	, mongoose = require('mongoose')
-	, UserModel = require('../lib/models/user').UserModel
+	, UserModel = require('../models/user').UserModel
     , fs = require('fs')
 	, crypto = require('crypto')
     , storage = require('./storage');
