+++ lib/stats.js
@@ -1,5 +1,5 @@
-var PackageModel = require('../lib/models/package').PackageModel
-	, UserModel = require('../lib/models/user').UserModel
+var PackageModel = require('../models/package').PackageModel
+	, UserModel = require('../models/user').UserModel
   , error = require('./error')
   , packages = require('./packages')
   , mongoose = require('mongoose')
