+++ lib/users.js
@@ -1,5 +1,5 @@
-var UserModel = require('../lib/models/user').UserModel
-	, PackageModel = require('../lib/models/package').PackageModel
+var UserModel = require('../models/user').UserModel
+	, PackageModel = require('../models/package').PackageModel
 	, error = require('./error')
 	, _ = require('underscore')
 	, async = require('async');
